fix init for multi downloads

+++ scripts/init.sh
@@ -18,11 +18,13 @@ fi
 
 # Check for DOCKER_ENABLE_SECURITY and download the appropriate JAR if required
 if [ "$DOCKER_ENABLE_SECURITY" = "true" ] && [ "$VERSION_TAG" != "alpha" ]; then
-	echo "Downloading from: https://github.com/Frooodle/Stirling-PDF/releases/download/$VERSION_TAG/Stirling-PDF-with-login.jar"
-    curl -L -o new-app.jar https://github.com/Frooodle/Stirling-PDF/releases/download/$VERSION_TAG/Stirling-PDF-with-login.jar
-    if [ $? -eq 0 ]; then  # checks if curl was successful
-        rm -f app.jar
-        mv new-app.jar app.jar
+    if [ ! -f app-security.jar ]; then
+        echo "Downloading from: https://github.com/Frooodle/Stirling-PDF/releases/download/$VERSION_TAG/Stirling-PDF-with-login.jar"
+        curl -L -o app-security.jar https://github.com/Frooodle/Stirling-PDF/releases/download/$VERSION_TAG/Stirling-PDF-with-login.jar
+        if [ $? -eq 0 ]; then  # checks if curl was successful
+            rm -f app.jar
+            ln -s app-security.jar app.jar
+        fi
     fi
 fi
